# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Parabellum POS API",
    "description": "Spécification complète et interactive des API du système de point de vente (POS) et de commande en ligne Gourmet. Supporte l'authentification par session, les webhooks tiers et les protocoles matériels réseau.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Serveur de développement local"
    },
    {
      "url": "https://www.parabellumpos.online",
      "description": "Serveur de production principal"
    }
  ],
  "paths": {
    "/api/orders": {
      "get": {
        "summary": "Liste des commandes du POS",
        "description": "Récupère la liste filtrée des commandes du restaurant. Accès réservé aux rôles CASHIER, SERVER, RESTAURATEUR, ADMIN.",
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "storeId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID du magasin (isolation multi-tenant)."
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/OrderStatus"
            },
            "description": "Filtrer par statut de commande."
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des commandes récupérée avec succès.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié."
          },
          "429": {
            "description": "Limite de requêtes dépassée (Rate limit)."
          }
        }
      },
      "post": {
        "summary": "Créer une commande en caisse (POS)",
        "description": "Crée une nouvelle commande sur place, à emporter ou en livraison depuis la caisse tactile.",
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCreateInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Commande créée avec succès.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Données d'entrée invalides (Zod error)."
          },
          "401": {
            "description": "Non authentifié."
          }
        }
      }
    },
    "/api/orders/{id}/status": {
      "post": {
        "summary": "Mettre à jour le statut d'une commande",
        "description": "Modifie le statut d'une commande et notifie en temps réel le KDS et l'interface client.",
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de la commande."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "$ref": "#/components/schemas/OrderStatus"
                  },
                  "prepTime": {
                    "type": "integer",
                    "description": "Temps estimé de préparation en minutes."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Statut mis à jour avec succès."
          },
          "400": {
            "description": "Statut invalide."
          },
          "404": {
            "description": "Commande non trouvée."
          }
        }
      }
    },
    "/api/orders/{id}/ticket": {
      "get": {
        "summary": "Payload d'impression du ticket de caisse",
        "description": "Récupère la structure d'impression brute (au format ESC/POS) pour impression physique du reçu de commande.",
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de la commande."
          }
        ],
        "responses": {
          "200": {
            "description": "Ticket généré avec succès.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketPayload"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/orders": {
      "post": {
        "summary": "Prise de commande externe",
        "description": "Soumet une commande externe (ex: depuis l'application mobile client ou le site web de commande).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicOrderInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Commande créée avec succès.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Erreur de validation des données d'entrée."
          }
        }
      }
    },
...
    "/api/webhooks/mobile-money": {
      "post": {
        "summary": "Webhook de notification de paiement Mobile Money",
        "tags": [
          "Webhooks"
        ],
        "description": "Point d'entrée de callback appelé par l'agrégateur (ex. CinetPay) pour notifier du statut d'une transaction Mobile Money (Orange Money, MTN, Moov).",
        "responses": {
          "200": {
            "description": "Notification reçue et validée avec succès.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "transaction_id manquant."
          },
          "401": {
            "description": "Signature HMAC invalide."
          },
          "505": {
            "description": "Erreur interne lors du traitement du webhook."
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token",
        "description": "Authentification par cookie NextAuth.js"
      },
      "GlovoSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Glovo-Signature",
        "description": "Signature HMAC cryptographique pour valider les requêtes de Glovo."
      }
    },
    "schemas": {
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "displayId": {
            "type": "string"
          },
          "storeId": {
            "type": "string"
          },
          "tableId": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          },
          "total": {
            "type": "number"
          },
          "tax": {
            "type": "number"
          },
          "discount": {
            "type": "number"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "OrderItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "productId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "quantity": {
            "type": "integer"
          },
          "options": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "OrderCreateInput": {
        "type": "object",
        "required": [
          "storeId",
          "items",
          "total"
        ],
        "properties": {
          "storeId": {
            "type": "string"
          },
          "tableId": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "productId",
                "quantity",
                "price"
              ],
              "properties": {
                "productId": {
                  "type": "string"
                },
                "quantity": {
                  "type": "integer"
                },
                "price": {
                  "type": "number"
                },
                "options": {
                  "type": "string"
                }
              }
            }
          },
          "total": {
            "type": "number"
          },
          "tax": {
            "type": "number"
          },
          "discount": {
            "type": "number"
          },
          "customerPhone": {
            "type": "string"
          }
        }
      },
      "PublicOrderInput": {
        "type": "object",
        "required": [
          "storeId",
          "items",
          "customerPhone"
        ],
        "properties": {
          "storeId": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "productId": {
                  "type": "string"
                },
                "quantity": {
                  "type": "integer"
                }
              }
            }
          },
          "customerPhone": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          }
        }
      },
...
      "TicketPayload": {
        "type": "object",
        "properties": {
          "header": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "footer": {
            "type": "string"
          },
          "rawEscPos": {
            "type": "string"
          }
        }
      },
      "OrderStatus": {
        "type": "string",
        "enum": [
          "EN_ATTENTE",
          "PREPARATION",
          "PRET",
          "COMPLETED",
          "CANCELLED"
        ],
        "description": "Status de la commande:\n- EN_ATTENTE: Commande reçue, en attente de préparation\n- PREPARATION: En cours de préparation en cuisine\n- PRET: Commande prête à être servie/remise\n- COMPLETED: Commande terminée et servie\n- CANCELLED: Commande annulée"
      },
...
    }
  },
...
}