# src/app/api/orders/[id]/ticket/route.ts
// src/app/api/orders/[id]/ticket/route.ts — Génération ticket PDF (format 80mm)

import { NextRequest, NextResponse } from "next/server";
import { getServerSession } from "next-auth/next";
import { authOptions, checkUserStoreAccess } from "@/lib/auth";
import { prisma } from "@/lib/db";

export const dynamic = "force-dynamic";

export async function GET(
  req: NextRequest,
  { params }: { params: Promise<{ id: string }> }
) {
  const session = await getServerSession(authOptions);
  if (!session?.user) {
    return new NextResponse("Authentification requise", { status: 401 });
  }

  const { id } = await params;

  try {
    const dbOrder = await prisma.order.findUnique({
      where: { id },
      include: {
        items: {
          include: {
            product: true,
          },
        },
        table: true,
      },
    });

    if (!dbOrder) {
      return new NextResponse("Commande introuvable", { status: 404 });
    }

    // Validation multi-tenant
    const hasAccess = await checkUserStoreAccess(session.user.id, session.user.role, dbOrder.storeId);
    if (!hasAccess) {
      return new NextResponse("Accès refusé à cet établissement", { status: 403 });
    }

    return generateTicketHTML({
      id:          dbOrder.id,
      orderNumber: `#${dbOrder.id.slice(-4).toUpperCase()}`,
      tableNumber: dbOrder.table?.number.toString() ?? null,
      type:        dbOrder.type,
      totalAmount: dbOrder.total,
      createdAt:   dbOrder.createdAt.toISOString(),
      items:       dbOrder.items.map((item) => ({
        name:      item.product.name,
        quantity:  item.quantity,
        unitPrice: item.price,
      })),
    });
  } catch (error: unknown) {
    const msg = error instanceof Error ? error.message : "Erreur interne du serveur";
    return new NextResponse(msg, { status: 500 });
  }
}

interface TicketData {
  id: string;
  orderNumber: string;
  tableNumber: string | null;
  type: string;
  totalAmount: number;
  createdAt: string;
  items: { name: string; quantity: number; unitPrice: number }[];
}

function generateTicketHTML(data: TicketData): NextResponse {
  const date = new Date(data.createdAt).toLocaleString("fr-FR", { dateStyle: "short", timeStyle: "short" });
  const typeLabel = { DINE_IN: "Sur place", TAKEAWAY: "À emporter", DELIVERY: "Livraison" }[data.type] ?? data.type;

  const itemsRows = data.items.map((item) =>
    `<tr>
      <td>${item.quantity}x ${item.name}</td>
      <td style="text-align:right">${(item.unitPrice * item.quantity).toFixed(2)} €</td>
    </tr>`
  ).join("");

  const html = `<!DOCTYPE html>
<html lang="fr">
<head>
  <meta charset="UTF-8">
  <title>Ticket ${data.orderNumber}</title>
  <style>
    @page { size: 80mm auto; margin: 4mm; }
    * { margin: 0; padding: 0; box-sizing: border-box; }
    body { font-family: 'Courier New', monospace; font-size: 12px; color: #000; width: 72mm; }
    .center { text-align: center; }
    .bold { font-weight: bold; }
    .big { font-size: 16px; font-weight: bold; }
    .divider { border-top: 1px dashed #000; margin: 6px 0; }
    table { width: 100%; border-collapse: collapse; }
    td { padding: 2px 0; }
    .total-row td { font-weight: bold; font-size: 14px; border-top: 1px solid #000; padding-top: 4px; }
    @media print {
      body { -webkit-print-color-adjust: exact; }
    }
  </style>
</head>
<body>
  <div class="center">
    <div class="big">🍔 Le Burger Doré</div>
    <div>12 Rue de Rivoli, 75001 Paris</div>
    <div>Tel: +33 1 42 00 10 20</div>
  </div>
  <div class="divider"></div>
  <div><span class="bold">Ticket:</span> ${data.orderNumber}</div>
  <div><span class="bold">Date:</span> ${date}</div>
  <div><span class="bold">Type:</span> ${typeLabel}</div>
  ${data.tableNumber ? `<div><span class="bold">Table:</span> ${data.tableNumber}</div>` : ""}
  <div class="divider"></div>
  <table>
    <tbody>${itemsRows}</tbody>
    <tfoot>
      <tr class="total-row">
        <td>TOTAL</td>
        <td style="text-align:right">${data.totalAmount.toFixed(2)} €</td>
      </tr>
    </tfoot>
  </table>
  <div class="divider"></div>
  <div class="center">Merci de votre visite !</div>
  <div class="center" style="font-size:10px; margin-top:4px">ID: ${data.id.slice(0, 8)}</div>
  <script>window.onload = () => window.print();</script>
</body>
</html>`;

  return new NextResponse(html, {
    headers: {
      "Content-Type":        "text/html; charset=utf-8",
      "Content-Disposition": `inline; filename="ticket-${data.orderNumber}.html"`,
    },
  });
}